Added support for different lighting models (chosen by the loaded game world) Slight improvement in memory useage with large worlds (much more is needed). Changed user settings so they can be put on different tabs. User settings are now saved/loaded from disk. Changed the way sleeping works in the chunk creation threads so they are used more during heavy loads. Updated version numbers in the assemblies. Added icons to executable files.

+++ TiVEPluginFramework/Lighting/PointLight.cs
@@ -4,7 +4,7 @@ namespace ProdigalSoftware.TiVEPluginFramework.Lighting
 {
     public sealed class PointLight : ILight
     {
-        public PointLight(Vector3b location, Color3f color, float lightBlockDist)
+        public PointLight(Vector3b location, Color3f color, int lightBlockDist)
         {
             Location = location;
             Color = color;
@@ -13,7 +13,7 @@ namespace ProdigalSoftware.TiVEPluginFramework.Lighting
 
         public Vector3b Location { get; private set; }
 
-        public float LightBlockDist { get; private set; }
+        public int LightBlockDist { get; private set; }
 
         public Color3f Color { get; private set; }
     }
